正在完善 HTTP API 以及 数据库接口

+++ ClusterDevicePlatform-client/src/main/java/cc/bitky/clusterdeviceplatform/client/netty/handler/ConfigHandler.java
@@ -37,6 +37,6 @@ public class ConfigHandler extends ChannelInboundHandlerAdapter {
         super.channelInactive(ctx);
         Attribute<Integer> key = ctx.channel().attr(AttributeKey.valueOf("ID"));
         logger.info("Channel「" + key.get() + "」已经断开连接");
-        tcpPresenter.channelInactive(key.get().intValue());
+        tcpPresenter.channelInactive(key.get());
     }
 }
